StockWebApi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.18
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
setuptools>=65.5.1
wheel>=0.38.4
